pyproject.toml
[project]
name = "finpulse"
version = "0.1.0"
description = "End-to-end pipeline: social/financial firehose -> S3 Parquet -> features -> signal"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.27",
    "pyarrow>=15",
    "duckdb>=1.0",
    "pandas>=2.2",
    "pydantic>=2.6",
    "structlog>=24.1",
    "python-dotenv>=1.0",
    "typer>=0.12",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "ruff>=0.4",
]

[project.scripts]
finpulse = "finpulse.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/finpulse"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "SIM"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q"
